Вынес события в singleton
