+++ services/EventService.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace app\services;
+
+use Yii;
+
+class EventService
+{
+    const EVENT_CREATED_TASK = 'created_task';
+    const EVENT_UPDATED_TASK = 'updated_task';
+    const EVENT_COMPLETED_TASK = 'completed_task';
+
+    const EVENT_INCOMING_SMS = 'incoming_sms';
+    const EVENT_OUTGOING_SMS = 'outgoing_sms';
+
+    const EVENT_INCOMING_CALL = 'incoming_call';
+    const EVENT_OUTGOING_CALL = 'outgoing_call';
+
+    const EVENT_INCOMING_FAX = 'incoming_fax';
+    const EVENT_OUTGOING_FAX = 'outgoing_fax';
+
+    const EVENT_CUSTOMER_CHANGE_TYPE = 'customer_change_type';
+    const EVENT_CUSTOMER_CHANGE_QUALITY = 'customer_change_quality';
+
+    /**
+     * @return array
+     */
+    public static function getEventTexts()
+    {
+        return [
+            self::EVENT_CREATED_TASK => Yii::t('app', 'Task created'),
+            self::EVENT_UPDATED_TASK => Yii::t('app', 'Task updated'),
+            self::EVENT_COMPLETED_TASK => Yii::t('app', 'Task completed'),
+
+            self::EVENT_INCOMING_SMS => Yii::t('app', 'Incoming message'),
+            self::EVENT_OUTGOING_SMS => Yii::t('app', 'Outgoing message'),
+
+            self::EVENT_CUSTOMER_CHANGE_TYPE => Yii::t('app', 'Type changed'),
+            self::EVENT_CUSTOMER_CHANGE_QUALITY => Yii::t('app', 'Property changed'),
+
+            self::EVENT_OUTGOING_CALL => Yii::t('app', 'Outgoing call'),
+            self::EVENT_INCOMING_CALL => Yii::t('app', 'Incoming call'),
+
+            self::EVENT_INCOMING_FAX => Yii::t('app', 'Incoming fax'),
+            self::EVENT_OUTGOING_FAX => Yii::t('app', 'Outgoing fax'),
+        ];
+    }
+}
